refactoring: splitting entities to multiple files
