+++ replay/scenarios/two_stages/two_stages_scenario.py
@@ -15,7 +15,8 @@ from replay.history_based_fp import HistoryBasedFeaturesProcessor
 from replay.metrics import Metric, Precision
 from replay.models import ALSWrap, RandomRec, PopRec
 from replay.models.base_rec import BaseRecommender, HybridRecommender
-from replay.scenarios.two_stages.reranker import LamaWrap, ReRankerModel, ReRanker
+from replay.scenarios.two_stages.lama_reranker import LamaWrap
+from replay.scenarios.two_stages.reranker import ReRankerModel, ReRanker
 from replay.scenarios.two_stages.slama_reranker import SlamaWrap
 from replay.session_handler import State
 from replay.splitters import Splitter, UserSplitter
